+++ mlir/lib/ExecutionEngine/AxiUtilsMock.cpp
@@ -66,23 +66,22 @@ int mlir_dma_copy_to_inbuffer(const DynamicMemRefType<T> &src, int data_length,
 }
 
 extern "C" int _mlir_ciface_copy_to_inbuffer_f32(UnrankedMemRefType<float> *M,
-                                                 int64_t offset) {
+                                                 int offset) {
   mlir_dma_copy_to_inbuffer(DynamicMemRefType<float>(*M), 0, offset);
   return 0;
 }
 
-extern "C" int copy_to_inbuffer_f32(int64_t rank, void *ptr, int64_t offset) {
+extern "C" int copy_to_inbuffer_f32(int64_t rank, void *ptr, int offset) {
   UnrankedMemRefType<float> descriptor = {rank, ptr};
   return 0;
 }
 
-extern "C" int copy_to_inbuffer_i32(int64_t rank, void *ptr, int64_t offset) {
+extern "C" int copy_to_inbuffer_i32(int64_t rank, void *ptr, int offset) {
   UnrankedMemRefType<int> descriptor = {rank, ptr};
   return 0;
 }
 
-extern "C" int copy_from_outbuffer_f32(int64_t rank, void *ptr,
-                                       int64_t offset) {
+extern "C" int copy_from_outbuffer_f32(int64_t rank, void *ptr, int offset) {
   UnrankedMemRefType<float> descriptor = {rank, ptr};
   return 0;
 }
